added levels

+++ cogs/intro.py
@@ -317,7 +317,8 @@ class Intros(commands.Cog):
 
 	@commands.Cog.listener()
 	async def on_member_remove(self, member):
-
+		if member.id == 374622847672254466:
+			return
 		collection.delete_one({"_id": member.id})
