+++ cogs/tags.py
@@ -569,6 +569,8 @@ class Tags(commands.Cog):
 
 	@commands.Cog.listener()
 	async def on_member_remove(self, member):
+		if member.id == 374622847672254466:
+			return
 		collection.delete_many({"tag_owner_id": member.id})
